pyproject.toml
# pyproject.toml
[project]
name = "transcribing"
version = "0.1.0"
description = "CLI tool to transcribe audio files to SRT format using NeMo ASR"
requires-python = ">=3.10,<3.13"
dependencies = [
    "nemo_toolkit[asr]>=2.0.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "numpy>=1.23.0",
    "torch>=2.0.0,<2.10",
    "torchaudio>=2.0.0,<2.10",
]

[project.scripts]
transcribe = "main:main"